I2C first fase
One of the two I2C functions is working, the one called "Read". The way of knowing it is that i requested the BQ34Z100 G1 the voltage that it's sensing by using the "read" function and the error 1% approximately.

Additionally i discovered that the address of the BQ34Z100 G1 is not 0xAA or 0xAB like the datasheet says, in fact it is 0x55 or 0xD5, the way of knowing that is that i used the function "TryCommunication", what it does is that it tries all the possible combinations of addresses that can by made out of one byte, when one address gets a response it shows it to you. It's important to notice that i tried using 0xAA to get the voltage and the result was not logical, the voltage that i was trying to measure was 3.99 V and the one that i got was 65.279 V...so, you can notice that it's no a logical value and assume that something is not correct.

i will be publishing more results in the near future.